simple implementation of ArrayList; ListTestsBase
no optimization when needing to resize
introduce ListTestsBase so that all of the tests can be shared with different implementations
