+++ UnitTests/ArrayListTests.java
@@ -1,0 +1,12 @@
+package UnitTests;
+
+import List.ArrayList;
+
+public class ArrayListTests extends ListTestsBase<ArrayList<Object>> {
+
+	@Override
+	protected ArrayList<Object> createInstance()
+	{
+		return new ArrayList<Object>();
+	}
+}
